+++ frontend/pages/botGame/botGame.html
@@ -11,41 +11,24 @@
 
 <body>
 
-    <button id="backButton" type="button"> 
+    <button id="backButton" type="button">
         <img src="../../resources/images/backArrow.png" alt="Back to homepage">
     </button>
-
-    <h2 id ="turn" class="center" > place your ships </h2>
-    <div id="gameBoardWrapper" class="flex">
-        <div class="gameBoard" id="leftGameBoard"></div>
-        <div class="gameBoard" id="rightGameBoard"></div>
-    </div>
-
-    <div class="shipSelectionWrapper center">
-        <div id="carrier" class="ship" draggable="true"></div>
-        <div id="battleship" class="ship" draggable="true"></div>
-        <div id="submarine" class="ship" draggable="true"></div>
-        <div id="cruiser" class="ship" draggable="true"></div>
-        <div id="destroyer" class="ship" draggable="true"></div>
-    </div>
-
-    <div class="functionWrapper">
-    
-        <div class="buttonWrapper">
-            <div id="gameButtonWrapper">
-                <input id="readyButton" type="button" value="Ready">
-            </div>
-            <div id="gameButtonWrapper">
-                <input id="randomizeButton" type="button" value="Randomize">
-            </div>
-            <div id="gameButtonWrapper">
-                <input id="resetButton" type="button" value="Reset">
-            </div>
+    <h2 id="turn" class="center"> Bot turn! </h2>
+    <div id="gameBoardWrapper" class="center">
+        <div>
+            <h2 id="yourName"> You </h2>
+            <div class="gameBoard" id="leftGameBoard"></div>
+        </div>
+        <div>
+            <h2 id="enemyName"> </h2>
+            <div class="gameBoard" id="rightGameBoard"></div>
         </div>
     </div>
+
     <script type="module" src="botGame.js"></script>
- 
-</div>
+
+    </div>
 
 </body>
