+++ frontend/utility/redirection.js
@@ -14,7 +14,7 @@ if (!User() || !User()._id || !User().name) {
 }
 
 // check for game status and redirect to right pages
-if (Game()) {
+if (Game() && Game().type !== "botGame") {
     // check for game in database
     await checkForGameInDB()
